+++ resources/views/hooks/body-end.blade.php
@@ -1,1 +1,1 @@
-@livewire(\Xuanpablo\FilamentPalette\Livewire\CommandPalette::class, ['panelId' => $panel?->getId()])
+@livewire(\Xuanpablo\CommandPalette\Livewire\CommandPalette::class, ['panelId' => $panel?->getId()])